+++ NEWS.md
@@ -1,4 +1,5 @@
 ### crunch 1.18.3 (under development)
+* Fix `bases()` when called on a univariate statistic (#124)
 
 ### crunch 1.18.2
 * `makeWeight` allows you to generate new weighting variables based on categorical variables (#80).